+++ packages/react-native-sdk/expo-config-plugin/src/withAndroidPermissions.ts
@@ -7,6 +7,7 @@ const withStreamVideoReactNativeSDKAndroidPermissions: ConfigPlugin = (
     'android.permission.POST_NOTIFICATIONS',
     'android.permission.FOREGROUND_SERVICE',
     'android.permission.FOREGROUND_SERVICE_MICROPHONE',
+    'android.permission.FOREGROUND_SERVICE_MEDIA_PROJECTION',
     'android.permission.BLUETOOTH',
     'android.permission.BLUETOOTH_CONNECT',
     'android.permission.BLUETOOTH_ADMIN',