+++ sample-apps/react/react-dogfood/lib/getServerSideCredentialsProps.ts
@@ -1,14 +1,20 @@
 import { GetServerSidePropsContext } from 'next';
 import { getServerSession } from 'next-auth';
 import { authOptions } from '../pages/api/auth/[...nextauth]';
-import { createToken } from '../helpers/jwt';
+import { createToken, decodeToken } from '../helpers/jwt';
 import type { User } from '@stream-io/video-react-sdk';
 
 export type ServerSideCredentialsProps = {
   user: User;
   userToken: string;
   apiKey: string;
-  gleapApiKey?: string;
+  gleapApiKey: string | null;
+};
+
+type QueryParams = {
+  api_key?: string;
+  token?: string;
+  user_id?: string;
 };
 
 export const getServerSideCredentialsProps = async (
@@ -24,11 +30,15 @@ export const getServerSideCredentialsProps = async (
     };
   }
 
-  const apiKey = process.env.STREAM_API_KEY as string;
+  const query = context.query as QueryParams;
+
+  const apiKey = query.api_key || (process.env.STREAM_API_KEY as string);
   const secretKey = process.env.STREAM_SECRET_KEY as string;
   const gleapApiKey = (process.env.GLEAP_API_KEY as string) || null;
 
-  const userIdOverride = context.query['user_id'] as string | undefined;
+  const userIdOverride = query.token
+    ? (decodeToken(query.token)['user_id'] as string | undefined)
+    : query.user_id;
   const userId = (
     userIdOverride ||
     session?.user?.name ||
@@ -38,16 +48,19 @@ export const getServerSideCredentialsProps = async (
   // Chat does not allow for Id's to include special characters
   const streamUserId = userId.replace(/[^_\-0-9a-zA-Z@]/g, '_');
   const userName = session.user?.name || userId;
+
+  const token = query.token || createToken(streamUserId, secretKey);
   return {
     props: {
       apiKey,
-      userToken: createToken(streamUserId, secretKey),
+      userToken: token,
       user: {
         id: streamUserId,
         name: userIdOverride || userName,
-        image: session.user?.image ?? null,
+        // @ts-expect-error - undefined is not serializable
+        image: session.user?.image || null,
       },
       gleapApiKey,
-    } as ServerSideCredentialsProps,
+    } satisfies ServerSideCredentialsProps,
   };
 };